+++ test/case-sort.sh
@@ -1,4 +1,6 @@
 
-./xsvutils ./test/data/sample-prefecture.tsv cut kana,name sort > ./var/test/case-sort-1.tsv
-./xsvutils ./test/data/sample-prefecture.tsv sort kana          > ./var/test/case-sort-2.tsv
+./xsvutils -v2 ./test/data/sort.tsv cut name,num sort > ./var/test/sort-1.tsv
+./xsvutils -v2 ./test/data/sort.tsv sort name         > ./var/test/sort-2.tsv
+./xsvutils -v2 ./test/data/sort.tsv sort num          > ./var/test/sort-3.tsv
+./xsvutils -v2 ./test/data/sort.tsv sort num:n        > ./var/test/sort-4.tsv