+++ CHANGELOG.md
@@ -12,6 +12,7 @@ To see compatibility information, `xsvutils help compatibility`.
 - remove `addlinenum2` subcommand
 - add `inscopy` subcommand on v2
 - change `sort` subcommand without parameters
+- remove `addnumsortable` subcommand
 
 ## version 0.2.2 (2018-03-07)
 - add `inshour` subcommand on v2